Refactoring Simulator v2.0; Adding BigInteger Conversion; Working

+++ src/main/java/com/sparta/engineering72/View/TextFileWriter.java
@@ -18,19 +18,19 @@ public class TextFileWriter {
         bufferedWriter.write("\n The Simulation ran for " + years + " years and " + months + " months");
         bufferedWriter.write("\n==============================================================");
         bufferedWriter.write("\n == RABBIT POPULATION REPORT == ");
-        bufferedWriter.write("\nThe total population of Rabbits: " + myFormatter.format(BigInteger.valueOf(reportPacker.getTotalRabbits())));
-        bufferedWriter.write("\nThe male population of Rabbits: " + myFormatter.format(BigInteger.valueOf(reportPacker.getMaleRabbits())));
-        bufferedWriter.write("\nThe female population of Rabbits: " + myFormatter.format(BigInteger.valueOf(reportPacker.getFemaleRabbits())));
+        bufferedWriter.write("\nThe total population of Rabbits: " + myFormatter.format(reportPacker.getTotalRabbits()));
+        bufferedWriter.write("\nThe male population of Rabbits: " + myFormatter.format(reportPacker.getMaleRabbits()));
+        bufferedWriter.write("\nThe female population of Rabbits: " + myFormatter.format(reportPacker.getFemaleRabbits()));
         bufferedWriter.write("\n--------------------------------------------------------------");
         bufferedWriter.write("\n == FOX POPULATION REPORT == ");
-        bufferedWriter.write("\nThe total population of Foxes: " + myFormatter.format(BigInteger.valueOf(reportPacker.getTotalFoxes())));
-        bufferedWriter.write("\nThe male population of Foxes: " + myFormatter.format(BigInteger.valueOf(reportPacker.getMaleFoxes())));
-        bufferedWriter.write("\nThe female population of Foxes: " + myFormatter.format(BigInteger.valueOf(reportPacker.getFemaleFoxes())));
+        bufferedWriter.write("\nThe total population of Foxes: " + myFormatter.format(reportPacker.getTotalFoxes()));
+        bufferedWriter.write("\nThe male population of Foxes: " + myFormatter.format(reportPacker.getMaleFoxes()));
+        bufferedWriter.write("\nThe female population of Foxes: " + myFormatter.format(reportPacker.getFemaleFoxes()));
         bufferedWriter.write("\n--------------------------------------------------------------");
         bufferedWriter.write("\n == DEATH REPORT == ");
-        bufferedWriter.write("\nThe total number of Rabbits that died: " + myFormatter.format(BigInteger.valueOf(reportPacker.getRabbitAgeDeaths())));
-        bufferedWriter.write("\nThe total number of Rabbits that were eaten: " + myFormatter.format(BigInteger.valueOf(reportPacker.getRabbitPreyDeaths())));
-        bufferedWriter.write("\nThe total number of Foxes that died: " + myFormatter.format(BigInteger.valueOf(reportPacker.getFoxDeaths())));
+        bufferedWriter.write("\nThe total number of Rabbits that died: " + myFormatter.format(reportPacker.getRabbitAgeDeaths()));
+        bufferedWriter.write("\nThe total number of Rabbits that were eaten: " + myFormatter.format(reportPacker.getRabbitPreyDeaths()));
+        bufferedWriter.write("\nThe total number of Foxes that died: " + myFormatter.format(reportPacker.getFoxDeaths()));
         bufferedWriter.write("\n==============================================================\n");
     }
